requirements.txt
channels
channels-redis
daphne
Django
django-cors-headers
django-filter
djangorestframework
drf-nested-routers
djangorestframework-simplejwt
djoser
drf-spectacular
Faker
gunicorn
Pillow
psycopg2-binary
python-dotenv
social-auth-app-django
django-jazzmin
tqdm
django-widget-tweaks